Moving x-fhir query resolver to quest
